fix structure; added gtihub
