+++ src/BodyTS.css
@@ -1,0 +1,149 @@
+.body__all {
+    display: flex;
+    max-width: 1440px;
+    margin: 0 auto;
+    flex-wrap: wrap;
+    gap: 20px;
+}
+
+.first__page {
+    display: flex;
+    height: 646px;
+}
+
+.back__Ground__firstPage {
+    display: flex;
+    background: linear-gradient(90deg, rgba(255, 255, 255, 0.9) 16.58%, rgba(255, 255, 255, 0) 72.38%), url(.jpg);
+}
+
+.bg_fp {
+    display: flex;
+    gap: 0 10px;
+}
+
+.slogan {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    row-gap: 20px;
+
+    color: #292E3D;
+
+    font-weight: 600;
+    font-size: 72px;
+    line-height: 84px;
+}
+
+.block_FP {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: flex-end;
+    justify-content: flex-end;
+    align-content: flex-end;
+    flex-direction: row;
+}
+
+.text__FP {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    flex-direction: row;
+
+    width: 416px;
+    height: 150px;
+
+    background: #3559C7;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 2px 2px 0px 0px;
+}
+
+.idk {
+    display: flex;
+    width: 163px;
+    height: 23px;
+    font-weight: 700;
+    font-size: 20px;
+    line-height: 23px;
+    text-align: center;
+    color: #FFFFFF;
+}
+
+.idk__2 {
+    display: flex;
+    width: 312px;
+    height: 61px;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 28px;
+    text-align: center;
+    color: #FFFFFF;
+}
+
+.arrow__Left {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    padding: 4px 20px;
+    gap: 10px;
+
+    width: 208px;
+    height: 40px;
+
+    background: #292E3D;
+    color: #FFFFFF;
+    cursor: pointer;
+}
+
+.arrow__Right {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    padding: 4px 20px;
+    gap: 10px;
+
+    width: 208px;
+    height: 40px;
+
+    background: #292E3D;
+    color: #FFFFFF;
+    cursor: pointer;
+}
+
+.img__Left {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+    text-align: center;
+}
+
+.img__Right {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+    text-align: center;
+}
+
+.arrows__andfml {
+    display: flex;
+    gap: 1px;
+}
+
+
+
+@media screen and (max-width: 1000px) {
+    .first__page {
+      flex-wrap: wrap;
+      row-gap: 20px;
+      flex: 1;
+    }
+}
+
+@media screen and (max-width: 320px) {
+    .first__page {
+      flex-wrap: wrap;
+      row-gap: 20px;
+      flex: 1;
+    }
+}
