+++ src/Navigation.css
@@ -1,0 +1,39 @@
+.Home {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+  
+    color: #14171F;
+  }
+  
+  .About {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+  
+    color: #14171F;
+  }
+  
+  .Projects {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+  
+    color: #14171F;
+  }
+  
+  .Services {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 21px;
+  
+    color: #14171F;
+  }
+  
+  .Contacts {
+    font-weight: 600;
+    font-size: 18px;
+    line-height: 21px;
+  
+    color: #F9995D;
+  }
