--- vite-project/src/index.css
@@ -1,13 +1,0 @@
-:root {
-    height: 5912px;
-    overflow-y: scroll;
-}
-
-body {
-    font-family: 'Work Sans', sans-serif;
-}
-
-a {
-    text-decoration: none;
-    color: inherit;
-}
